pyproject.toml
[project]
name = "potato-achis"
version = "0.1.0"
description = "Multi-source Domain Adaptation for Andean Potato Disease Classification"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "R0SEWT", email = "your-email@example.com"}
]
keywords = ["deep-learning", "domain-adaptation", "plant-disease", "pytorch"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "timm>=0.9.0",
    "numpy>=1.24.0",
    "pillow>=9.0.0",
    "matplotlib>=3.7.0",
    "scikit-learn>=1.2.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
viz = [
    "seaborn>=0.12.0",
    "torchcam>=0.3.0",
]
tracking = [
    "wandb>=0.15.0",
    "tensorboard>=2.13.0",
]
all = [
    "potato-achis[dev,viz,tracking]",
]

[project.scripts]
potato-train = "src.train:main"
potato-eval = "src.eval:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
skip_gitignore = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
